requirements.txt
absl-py==2.0.0
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
blinker==1.6.3
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
comm==0.1.4
contourpy==1.1.1
cvzone==1.6.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
dlib==19.24.2
executing==2.0.0
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.43.1
gTTS==2.4.0
idna==3.4
imutils==0.5.4
ipykernel==6.25.2
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.3.1
jupyter_core==5.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mediapipe==0.10.7
nest-asyncio==1.5.8
numpy==1.26.0
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
playsound==1.3.0
print-color==0.4.6
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pygame==2.5.2
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
pyzmq==25.1.1
regex==2023.10.3
requests==2.31.0
scipy==1.11.3
screen-brightness-control==0.22.1
six==1.16.0
sounddevice==0.4.6
soundfile==0.12.1
stack-data==0.6.3
timer==0.2.2
timer.py==2.4.1
tk==0.1.0
tornado==6.3.3
traitlets==5.11.2
urllib3==2.0.6
wcwidth==0.2.8
Werkzeug==3.0.0
WMI==1.4.9
